+++ CHANGELOG.md
@@ -15,6 +15,10 @@ You can find our backwards-compatibility policy [here](https://github.com/hynek/
 
 ## [Unreleased](https://github.com/hynek/structlog/compare/24.1.0...HEAD)
 
+### Changed
+
+- `structlog.processors.LogfmtRenderer` now escapes newlines.
+  [#592](https://github.com/hynek/structlog/pull/592)
 
 ## [24.1.0](https://github.com/hynek/structlog/compare/23.3.0...24.1.0) - 2024-01-08